document-service/requirements.txt
# =============================================================================
# PayLockr Document Processing Service — Python Dependencies
# =============================================================================

# --- Web framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9    # Required for form/file uploads in FastAPI

# --- HTTP client (for fetching files from URLs / Supabase Storage) ---
httpx>=0.27.0

# --- PDF text extraction ---
pdfplumber>=0.11.0
PyMuPDF>=1.24.0            # also provides scanned PDF → image conversion (ghostscript-free)

# --- Table extraction from text PDFs ---
camelot-py[cv]>=0.11.0     # requires ghostscript + OpenCV

# --- Image preprocessing ---
opencv-python-headless>=4.9.0   # headless: no GUI deps
Pillow>=10.3.0

# --- OCR ---
pytesseract>=0.3.10        # Python wrapper; Tesseract binary must be installed separately

# --- Environment / config ---
python-dotenv>=1.0.1

# --- Optional: LLM normalization / Vision (Gemini) ---
google-genai>=1.0.0              # new official Google GenAI SDK (replaces google-generativeai)
